pyproject.toml
[global]
extra-index-url = "https://www.piwheels.org/simple"

[tool.poetry]
name = "enviroweb"
version = "0.1.0"
description = ""
authors = ["xNinjaKittyx"]

[tool.poetry.dependencies]
python = "^3.7"
pms5003 = "^0.0.5"
"RPi.GPIO" = "^0.7.0"
ltr559 = "^0.1.1"
smbus2 = "^0.4.0"
ST7735 = "^0.0.4"
numpy = "^1.19.4"
enviroplus = "^0.0.3"
pimoroni-bme280 = "^0.1.1"
fastapi = "^0.62.0"
uvicorn = "^0.12.3"
psutil = "^5.7.3"
Pillow = "^8.0.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
